распределила функции по модулям
